Strip back and rewire Portfolio
- Wire portfolio into ExecutionEngine
- Remove AccountInquiry
- Standardize blank lines in class docstrings
