+++ nautilus_trader/trading/portfolio.pxd
@@ -18,13 +18,11 @@ from cpython.datetime cimport date
 from nautilus_trader.common.clock cimport Clock
 from nautilus_trader.common.logging cimport LoggerAdapter
 from nautilus_trader.common.uuid cimport UUIDFactory
-from nautilus_trader.model.currency cimport Currency
 from nautilus_trader.model.events cimport PositionClosed
 from nautilus_trader.model.events cimport PositionEvent
 from nautilus_trader.model.events cimport PositionModified
 from nautilus_trader.model.events cimport PositionOpened
-from nautilus_trader.model.identifiers cimport Symbol
-from nautilus_trader.model.objects cimport Money
+from nautilus_trader.trading.account cimport Account
 
 
 cdef class Portfolio:
@@ -32,23 +30,15 @@ cdef class Portfolio:
     cdef Clock _clock
     cdef UUIDFactory _uuid_factory
 
+    cdef dict _accounts
     cdef dict _positions_open
     cdef dict _positions_closed
 
     cdef readonly date date_now
-    cdef readonly Currency base_currency
-    cdef readonly Money daily_pnl_realized
-    cdef readonly Money total_pnl_realized
 
-    cpdef void set_base_currency(self, Currency currency) except *
+    cpdef void register_account(self, Account account) except *
     cpdef void update(self, PositionEvent event) except *
     cpdef void reset(self) except *
-    cpdef set symbols_open(self)
-    cpdef set symbols_closed(self)
-    cpdef set symbols_all(self)
-    cpdef dict positions_open(self, Symbol symbol=*)
-    cpdef dict positions_closed(self, Symbol symbol=*)
-    cpdef dict positions_all(self, Symbol symbol=*)
 
     cdef void _handle_position_opened(self, PositionOpened event) except *
     cdef void _handle_position_modified(self, PositionModified event) except *
