+++ tests/unit_tests/model/test_model_identifiers.py
@@ -19,6 +19,7 @@ from nautilus_trader.core.types import Identifier
 from nautilus_trader.model.enums import AccountType
 from nautilus_trader.model.identifiers import AccountId
 from nautilus_trader.model.identifiers import ClientOrderId
+from nautilus_trader.model.identifiers import Issuer
 from nautilus_trader.model.identifiers import PositionId
 from nautilus_trader.model.identifiers import StrategyId
 from nautilus_trader.model.identifiers import Symbol
@@ -142,5 +143,5 @@ class IdentifierTests(unittest.TestCase):
         self.assertEqual(account_id1, account_id1)
         self.assertNotEqual(account_id1, account_id2)
         self.assertEqual("FXCM-02851908-DEMO", account_id1.value)
-        self.assertEqual("FXCM", account_id1.issuer)
+        self.assertEqual(Issuer("FXCM"), account_id1.issuer)
         self.assertEqual(account_id1, AccountId.py_from_string("FXCM-02851908-DEMO"))
